backend/requirements.txt
# Generated by: uv pip compile requirements.in -o requirements.txt
# Do not edit manually — edit requirements.in and run: make compile

langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
respx>=0.21.0
